XWIKI-15912: Improper "rename log" string is displayed when user is performing a page copy
  * Start to deprecate and move to legacy the create API from
RefactoringScriptService
  * Introduce the new RequestFactory
  * Introduce the new CopyRequest
