+++ xwiki-platform-core/xwiki-platform-legacy/pom.xml
@@ -50,6 +50,7 @@
     <module>xwiki-platform-legacy-url</module>
     <module>xwiki-platform-legacy-web</module>
     <module>xwiki-platform-legacy-watchlist-api</module>
+    <module>xwiki-platform-legacy-refactoring</module>
   </modules>
   <build>
     <pluginManagement>
