+++ xwiki-platform-core/xwiki-platform-refactoring/xwiki-platform-refactoring-api/src/main/java/org/xwiki/refactoring/script/RefactoringScriptService.java
@@ -33,6 +33,8 @@ import javax.inject.Singleton;
 import org.apache.commons.lang3.StringUtils;
 import org.xwiki.bridge.DocumentAccessBridge;
 import org.xwiki.component.annotation.Component;
+import org.xwiki.component.phase.Initializable;
+import org.xwiki.component.phase.InitializationException;
 import org.xwiki.context.Execution;
 import org.xwiki.job.Job;
 import org.xwiki.job.JobException;
@@ -44,8 +46,7 @@ import org.xwiki.model.reference.DocumentReference;
 import org.xwiki.model.reference.EntityReference;
 import org.xwiki.model.reference.EntityReferenceProvider;
 import org.xwiki.model.reference.SpaceReference;
-import org.xwiki.model.reference.WikiReference;
-import org.xwiki.refactoring.job.AbstractDeletedDocumentsRequest;
+import org.xwiki.refactoring.job.CopyRequest;
 import org.xwiki.refactoring.job.CreateRequest;
 import org.xwiki.refactoring.job.EntityRequest;
 import org.xwiki.refactoring.job.MoveRequest;
@@ -65,7 +66,7 @@ import org.xwiki.security.authorization.Right;
 @Component
 @Named(RefactoringJobs.GROUP)
 @Singleton
-public class RefactoringScriptService implements ScriptService
+public class RefactoringScriptService implements ScriptService, Initializable
 {
     /**
      * The key under which the last encountered error is stored in the current execution context.
@@ -102,181 +103,21 @@ public class RefactoringScriptService implements ScriptService
     @Inject
     private ModelContext modelContext;
 
-    /**
-     * Creates a request to move the specified source entities to the specified destination entity (which becomes their
-     * new parent).
-     *
-     * @param sources specifies the entities to be moved
-     * @param destination specifies the place where to move the entities (their new parent entity)
-     * @return the move request
-     */
-    public MoveRequest createMoveRequest(Collection<EntityReference> sources, EntityReference destination)
+    private RequestFactory requestFactory;
+
+    @Override
+    public void initialize() throws InitializationException
     {
-        return createMoveRequest(RefactoringJobs.MOVE, sources, destination);
+        this.requestFactory = new RequestFactory(documentAccessBridge, modelContext);
     }
 
     /**
-     * Creates a request to move the specified source entity to the specified destination entity (which becomes its new
-     * parent).
-     *
-     * @param source specifies the entity to be moved
-     * @param destination specifies the place where to move the source entity (its new parent entity)
-     * @return the move request
+     * @return the standard request factory for creating the different refactoring requests.
+     * @since 10.11RC1
      */
-    public MoveRequest createMoveRequest(EntityReference source, EntityReference destination)
+    public RequestFactory getRequestFactory()
     {
-        return createMoveRequest(Arrays.asList(source), destination);
-    }
-
-    /**
-     * Creates a request to rename the entity specified by the given old reference.
-     *
-     * @param oldReference the entity to rename
-     * @param newReference the new entity reference after the rename
-     * @return the rename request
-     */
-    public MoveRequest createRenameRequest(EntityReference oldReference, EntityReference newReference)
-    {
-        return createMoveRequest(RefactoringJobs.RENAME, Collections.singletonList(oldReference), newReference);
-    }
-
-    /**
-     * Creates a request to rename the specified entity.
-     *
-     * @param reference the entity to rename
-     * @param newName the new entity name
-     * @return the rename request
-     */
-    public MoveRequest createRenameRequest(EntityReference reference, String newName)
-    {
-        return createRenameRequest(reference, new EntityReference(newName, reference.getType(), reference.getParent()));
-    }
-
-    /**
-     * Creates a request to copy the specified source entities to the specified destination entity.
-     *
-     * @param sources specifies the entities to be copied
-     * @param destination specifies the place where to copy the entities (becomes the parent of the copies)
-     * @return the copy request
-     */
-    public MoveRequest createCopyRequest(Collection<EntityReference> sources, EntityReference destination)
-    {
-        MoveRequest request = createMoveRequest(RefactoringJobs.COPY, sources, destination);
-        request.setDeleteSource(false);
-        request.setAutoRedirect(false);
-        request.setUpdateParentField(false);
-        return request;
-    }
-
-    /**
-     * Creates a request to copy the specified source entity to the specified destination entity.
-     *
-     * @param source specifies the entity to be copied
-     * @param destination specifies the place where to copy the source entity (becomes the parent of the copy)
-     * @return the copy request
-     */
-    public MoveRequest createCopyRequest(EntityReference source, EntityReference destination)
-    {
-        return createCopyRequest(Arrays.asList(source), destination);
-    }
-
-    /**
-     * Creates a request to copy the specified entity with a different reference.
-     *
-     * @param sourceReference the entity to copy
-     * @param copyReference the reference to use for the copy
-     * @return the copy-as request
-     */
-    public MoveRequest createCopyAsRequest(EntityReference sourceReference, EntityReference copyReference)
-    {
-        MoveRequest request = createMoveRequest(RefactoringJobs.COPY_AS, Arrays.asList(sourceReference), copyReference);
-        request.setDeleteSource(false);
-        request.setAutoRedirect(false);
-        request.setUpdateParentField(false);
-        return request;
-    }
-
-    /**
-     * Creates a request to copy the specified entity with a different name.
-     *
-     * @param reference the entity to copy
-     * @param copyName the name of the entity copy
-     * @return the copy-as request
-     */
-    public MoveRequest createCopyAsRequest(EntityReference reference, String copyName)
-    {
-        EntityReference copyReference = new EntityReference(copyName, reference.getType(), reference.getParent());
-        return createCopyAsRequest(reference, copyReference);
-    }
-
-    /**
-     * Creates a request to delete the specified entities.
-     *
-     * @param entityReferences the entities to delete
-     * @return the delete request
-     */
-    public EntityRequest createDeleteRequest(Collection<EntityReference> entityReferences)
-    {
-        EntityRequest request = new EntityRequest();
-        initEntityRequest(request, RefactoringJobs.DELETE, entityReferences);
-        return request;
-    }
-
-    /**
-     * Creates a request to create the specified entities.
-     *
-     * @param entityReferences the entities to create
-     * @return the create request
-     * @since 7.4M2
-     */
-    public CreateRequest createCreateRequest(Collection<EntityReference> entityReferences)
-    {
-        CreateRequest request = new CreateRequest();
-        initEntityRequest(request, RefactoringJobs.CREATE, entityReferences);
-        // Set deep create by default, to copy (if possible) any existing hierarchy of a specified template document.
-        // TODO: expose this in the create UI to advanced users to allow them to opt-out?
-        request.setDeep(true);
-        return request;
-    }
-
-    private void initEntityRequest(EntityRequest request, String type, Collection<EntityReference> entityReferences)
-    {
-        request.setId(generateJobId(type));
-        request.setJobType(type);
-        request.setEntityReferences(entityReferences);
-        setRightsProperties(request);
-    }
-
-    private MoveRequest createMoveRequest(String type, Collection<EntityReference> sources, EntityReference destination)
-    {
-        MoveRequest request = new MoveRequest();
-        initEntityRequest(request, type, sources);
-        request.setDestination(destination);
-        request.setUpdateLinks(true);
-        request.setAutoRedirect(true);
-        request.setUpdateParentField(true);
-        return request;
-    }
-
-    /**
-     * @param type the type of refactoring
-     * @return an id for a job to perform the specified type of refactoring
-     */
-    private List<String> generateJobId(String type)
-    {
-        String suffix = new Date().getTime() + "-" + ThreadLocalRandom.current().nextInt(100, 1000);
-        return getJobId(type, suffix);
-    }
-
-    /**
-     * @param type the type of refactoring
-     * @param suffix uniquely identifies the job among those of the specified type
-     * @return an id for a job to perform the specified type of refactoring
-     */
-    private List<String> getJobId(String type, String suffix)
-    {
-        return Arrays
-            .asList(RefactoringJobs.GROUP, StringUtils.removeStart(type, RefactoringJobs.GROUP_PREFIX), suffix);
+        return this.requestFactory;
     }
 
     /**
@@ -302,7 +143,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job move(Collection<EntityReference> sources, EntityReference destination)
     {
-        return move(createMoveRequest(sources, destination));
+        return move(getRequestFactory().createMoveRequest(sources, destination));
     }
 
     /**
@@ -316,7 +157,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job move(EntityReference source, EntityReference destination)
     {
-        return move(createMoveRequest(source, destination));
+        return move(getRequestFactory().createMoveRequest(source, destination));
     }
 
     /**
@@ -341,7 +182,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job rename(EntityReference oldReference, EntityReference newReference)
     {
-        return rename(createRenameRequest(oldReference, newReference));
+        return rename(getRequestFactory().createRenameRequest(oldReference, newReference));
     }
 
     /**
@@ -354,7 +195,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job rename(EntityReference reference, String newName)
     {
-        return rename(createRenameRequest(reference, newName));
+        return rename(getRequestFactory().createRenameRequest(reference, newName));
     }
 
     /**
@@ -410,7 +251,22 @@ public class RefactoringScriptService implements ScriptService
      * @param request the copy request to perform
      * @return the job that has been scheduled and that can be used to monitor the progress of the operation,
      *         {@code null} in case of failure
+     * @since 10.11RC1
      */
+    public Job copy(CopyRequest request)
+    {
+        return execute(RefactoringJobs.COPY, request);
+    }
+
+    /**
+     * Schedules an asynchronous job to perform the given copy request.
+     *
+     * @param request the copy request to perform
+     * @return the job that has been scheduled and that can be used to monitor the progress of the operation,
+     *         {@code null} in case of failure
+     * @deprecated Use {@link #copy(CopyRequest)} instead.
+     */
+    @Deprecated
     public Job copy(MoveRequest request)
     {
         return execute(RefactoringJobs.COPY, request);
@@ -426,7 +282,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job copy(Collection<EntityReference> sources, EntityReference destination)
     {
-        return copy(createCopyRequest(sources, destination));
+        return copy(getRequestFactory().createCopyRequest(sources, destination));
     }
 
     /**
@@ -439,7 +295,19 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job copy(EntityReference source, EntityReference destination)
     {
-        return copy(createCopyRequest(source, destination));
+        return copy(getRequestFactory().createCopyRequest(source, destination));
+    }
+
+    /**
+     * Schedules an asynchronous job to perform the given copy-as request.
+     *
+     * @param request the copy-as request to perform
+     * @return the job that has been scheduled and that can be used to monitor the progress of the operation,
+     *         {@code null} in case of failure
+     */
+    public Job copyAs(CopyRequest request)
+    {
+        return execute(RefactoringJobs.COPY, request);
     }
 
     /**
@@ -464,7 +332,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job copyAs(EntityReference sourceReference, EntityReference copyReference)
     {
-        return copyAs(createCopyAsRequest(sourceReference, copyReference));
+        return copyAs(getRequestFactory().createCopyAsRequest(sourceReference, copyReference));
     }
 
     /**
@@ -477,7 +345,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job copyAs(EntityReference reference, String copyName)
     {
-        return copyAs(createCopyAsRequest(reference, copyName));
+        return copyAs(getRequestFactory().createCopyAsRequest(reference, copyName));
     }
 
     /**
@@ -501,7 +369,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job delete(Collection<EntityReference> entityReferences)
     {
-        return delete(createDeleteRequest(entityReferences));
+        return delete(getRequestFactory().createDeleteRequest(entityReferences));
     }
 
     /**
@@ -539,7 +407,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job create(Collection<EntityReference> entityReferences)
     {
-        return create(createCreateRequest(entityReferences));
+        return create(getRequestFactory().createCreateRequest(entityReferences));
     }
 
     /**
@@ -569,7 +437,7 @@ public class RefactoringScriptService implements ScriptService
 
         // Make sure that only the PR users can change the rights and context properties from the request.
         if (!this.authorization.hasAccess(Right.PROGRAM)) {
-            setRightsProperties(request);
+            getRequestFactory().setRightsProperties(request);
         }
 
         try {
@@ -578,13 +446,6 @@ public class RefactoringScriptService implements ScriptService
             setError(e);
             return null;
         }
-    }
-
-    private void setRightsProperties(AbstractCheckRightsRequest request)
-    {
-        request.setCheckRights(true);
-        request.setUserReference(this.documentAccessBridge.getCurrentUserReference());
-        request.setAuthorReference(this.documentAccessBridge.getCurrentAuthorReference());
     }
 
     /**
@@ -606,62 +467,6 @@ public class RefactoringScriptService implements ScriptService
     private void setError(Exception e)
     {
         this.execution.getContext().setProperty(REFACTORING_ERROR_KEY, e);
-    }
-
-    /**
-     * Creates a request to restore a specified batch of deleted documents from the recycle bin.
-     *
-     * @param batchId the ID of the batch of deleted documents to restore
-     * @return the restore request
-     * @since 9.4RC1
-     */
-    public RestoreRequest createRestoreRequest(String batchId)
-    {
-        RestoreRequest request = initializeRestoreRequest();
-        request.setBatchId(batchId);
-        return request;
-    }
-
-    /**
-     * Creates a request to restore a specified list of deleted documents from the recycle bin.
-     *
-     * @param deletedDocumentIds the list of IDs of the deleted documents to restore
-     * @return the restore request
-     * @since 9.4RC1
-     */
-    public RestoreRequest createRestoreRequest(List<Long> deletedDocumentIds)
-    {
-        RestoreRequest request = initializeRestoreRequest();
-        request.setDeletedDocumentIds(deletedDocumentIds);
-        return request;
-    }
-
-    private RestoreRequest initializeRestoreRequest()
-    {
-        RestoreRequest request = new RestoreRequest();
-
-        request.setId(generateJobId(RefactoringJobs.RESTORE));
-        request.setCheckRights(true);
-        request.setUserReference(this.documentAccessBridge.getCurrentUserReference());
-        request.setWikiReference(getCurrentWikiReference());
-
-        return request;
-    }
-
-    private WikiReference getCurrentWikiReference()
-    {
-        WikiReference result = null;
-
-        EntityReference currentEntityReference = this.modelContext.getCurrentEntityReference();
-        if (currentEntityReference != null) {
-            EntityReference wikiEntityReference =
-                this.modelContext.getCurrentEntityReference().extractReference(EntityType.WIKI);
-            if (wikiEntityReference != null) {
-                result = new WikiReference(wikiEntityReference);
-            }
-        }
-
-        return result;
     }
 
     /**
@@ -687,7 +492,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job restore(String batchId)
     {
-        return this.restore(this.createRestoreRequest(batchId));
+        return this.restore(getRequestFactory().createRestoreRequest(batchId));
     }
 
     /**
@@ -700,47 +505,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job restore(List<Long> deletedDocumentIds)
     {
-        return this.restore(this.createRestoreRequest(deletedDocumentIds));
-    }
-
-    /**
-     * Creates a request to permanently delete a specified batch of deleted documents from the recycle bin.
-     *
-     * @param batchId the ID of the batch of deleted documents to permanently delete
-     * @return the permanently delete request
-     * @since 10.10RC1
-     */
-    public PermanentlyDeleteRequest createPermanentlyDeleteRequest(String batchId)
-    {
-        PermanentlyDeleteRequest request = initializePermanentlyDeleteRequest();
-        request.setBatchId(batchId);
-        return request;
-    }
-
-    /**
-     * Creates a request to permanently delete a specified list of deleted documents from the recycle bin.
-     *
-     * @param deletedDocumentIds the list of IDs of the deleted documents to permanently delete
-     * @return the permanently delete request
-     * @since 10.10RC1
-     */
-    public PermanentlyDeleteRequest createPermanentlyDeleteRequest(List<Long> deletedDocumentIds)
-    {
-        PermanentlyDeleteRequest request = initializePermanentlyDeleteRequest();
-        request.setDeletedDocumentIds(deletedDocumentIds);
-        return request;
-    }
-
-    private PermanentlyDeleteRequest initializePermanentlyDeleteRequest()
-    {
-        PermanentlyDeleteRequest request = new PermanentlyDeleteRequest();
-
-        request.setId(generateJobId(RefactoringJobs.PERMANENTLY_DELETE));
-        request.setCheckRights(true);
-        request.setUserReference(this.documentAccessBridge.getCurrentUserReference());
-        request.setWikiReference(getCurrentWikiReference());
-
-        return request;
+        return this.restore(getRequestFactory().createRestoreRequest(deletedDocumentIds));
     }
 
     /**
@@ -766,7 +531,7 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job permanentlyDelete(String batchId)
     {
-        return this.permanentlyDelete(this.createPermanentlyDeleteRequest(batchId));
+        return this.permanentlyDelete(getRequestFactory().createPermanentlyDeleteRequest(batchId));
     }
 
     /**
@@ -780,6 +545,6 @@ public class RefactoringScriptService implements ScriptService
      */
     public Job permanentlyDelete(List<Long> deletedDocumentIds)
     {
-        return this.permanentlyDelete(this.createPermanentlyDeleteRequest(deletedDocumentIds));
+        return this.permanentlyDelete(getRequestFactory().createPermanentlyDeleteRequest(deletedDocumentIds));
     }
 }
