+++ xwiki-platform-core/xwiki-platform-web/src/main/webapp/templates/rename.vm
@@ -53,7 +53,8 @@
             [${escapetool.xml($newDocReferenceAsString)}]))
           #template('renameStep1.vm')
         #else
-          #set ($renameRequest = $services.refactoring.createRenameRequest($doc.documentReference, $newDocReference))
+          #set ($renameRequest = $services.refactoring.requestFactory.createRenameRequest($doc.documentReference,
+            $newDocReference))
           ## We can't rename the children if the new page is terminal.
           #set ($deep = $request.deep == 'true' && !$terminal)
           #set ($discard = $renameRequest.setDeep($deep))
