+++ xwiki-platform-distribution/xwiki-platform-distribution-war-legacydependencies/pom.xml
@@ -236,5 +236,10 @@
         </exclusion>
       </exclusions>
     </dependency>
+    <dependency>
+      <groupId>org.xwiki.platform</groupId>
+      <artifactId>xwiki-platform-legacy-refactoring-api</artifactId>
+      <version>${project.version}</version>
+    </dependency>
   </dependencies>
 </project>
